refactor(2024): day 13 with ksmt

+++ advent2024/src/test/kotlin/year2024/Day13Test.kt
@@ -10,5 +10,5 @@ class Day13Test {
     fun part1() = assertEquals(480L, day.testPart1())
 
     @Test
-    fun part2() = assertEquals((875318608908L).toBigInteger(), day.testPart2())
+    fun part2() = assertEquals((875318608908L), day.testPart2())
 }
